Move some ert_config tests to lark_parse tests
